WIP adding directionsRequest
